feedback changes & split some MovementState stuff out into separate files
