+++ Maple2.Server.Game/Model/Field/Actor/ActorStateComponent/MovementState.cs
@@ -32,26 +32,6 @@ public partial class MovementState {
     private float baseSpeed = 0;
     private readonly float aniSpeed = 1;
 
-    #region WalkProperties
-    private enum WalkType {
-        None,
-        Direction,
-        MoveTo,
-        ToTarget,
-        FromTarget
-    }
-    private Vector3 walkDirection;
-    private Vector3 walkTargetPosition;
-    private float walkTargetDistance;
-    private WalkType walkType = WalkType.None;
-    private (Vector3 start, Vector3 end) walkSegment;
-    private bool walkSegmentSet;
-    private bool walkLookWhenDone = false;
-    private AnimationSequence? walkSequence = null;
-    private float walkSpeed;
-    private NpcTask? walkTask = null;
-    #endregion
-
     #region CastSkill
     private SkillRecord? castSkill = null;
     private NpcTask? castTask = null;
@@ -59,24 +39,6 @@ public partial class MovementState {
 
     #region Emote
     private NpcTask? emoteActionTask = null;
-    #endregion
-
-    #region Debugging
-    private class DebugMarker {
-        public bool Spawned;
-        public Item ItemData;
-        public FieldItem Item;
-        public long NextUpdate = 0;
-        public int ObjectId = 0;
-
-        public DebugMarker(FieldItem item, Item itemData) {
-            Item = item;
-            ItemData = itemData;
-        }
-    }
-    private DebugMarker debugNpc;
-    private DebugMarker debugTarget;
-    private DebugMarker debugAgent;
     #endregion
 
     public MovementState(FieldNpc actor) {
@@ -92,47 +54,6 @@ public partial class MovementState {
         debugNpc = InitDebugMarker(30000071, 2);
         debugTarget = InitDebugMarker(50300014, 6);
         debugAgent = InitDebugMarker(30000084, 3);
-    }
-
-    private void UpdateMoveSpeed(float speed) {
-        Stat moveSpeed = actor.Stats[BasicAttribute.MovementSpeed];
-
-        speedOverride = speed;
-        Speed = speed == 0 ? (float) moveSpeed.Current / 100 : speed;
-        Speed *= baseSpeed;
-    }
-
-    private void StartWalking(string sequence, NpcTask task) {
-        sequence = sequence == "" ? "Run_A" : sequence;
-        walkSegmentSet = false;
-        walkSpeed = Speed;
-
-        emoteActionTask?.Cancel();
-
-        bool isWalking = sequence.StartsWith("Walk_");
-
-        baseSpeed = isWalking ? actor.Value.Metadata.Action.WalkSpeed : actor.Value.Metadata.Action.RunSpeed;
-
-        if (actor.AnimationState.PlayingSequence?.Name == sequence || actor.AnimationState.TryPlaySequence(sequence, aniSpeed * Speed, AnimationType.Misc)) {
-            stateSequence = actor.AnimationState.PlayingSequence;
-            walkSequence = stateSequence;
-            walkTask = task;
-
-            SetState(ActorState.Walk, isWalking ? ActorSubState.Walk_Walking : ActorSubState.Walk_Running);
-        } else {
-            task.Cancel();
-
-            Idle();
-        }
-    }
-
-    public bool IsMovingToTarget() {
-        return State == ActorState.Walk && walkType switch {
-            WalkType.MoveTo => true,
-            WalkType.FromTarget => true,
-            WalkType.ToTarget => true,
-            _ => false
-        };
     }
 
     public NpcTask TryMoveDirection(Vector3 direction, bool isBattle, string sequence = "", float speed = 0) {
@@ -249,24 +170,6 @@ public partial class MovementState {
         }
     }
 
-    public void StateWalkEvent(string keyName) {
-        switch (keyName) {
-            case "end":
-                if (State == ActorState.Walk) {
-                    if (actor.AnimationState.TryPlaySequence(stateSequence!.Name, aniSpeed * Speed, AnimationType.Misc)) {
-                        stateSequence = actor.AnimationState.PlayingSequence;
-                    }
-
-                    return;
-                }
-                walkTask?.Cancel();
-
-                break;
-            default:
-                break;
-        }
-    }
-
     public void StateSkillEvent(string keyName) {
         switch (keyName) {
             case "end":
@@ -310,99 +213,6 @@ public partial class MovementState {
                 break;
             default:
                 break;
-        }
-    }
-
-    private void StateWalkUpdate(long tickCount, long tickDelta) {
-        if (actor.Navigation is null) {
-            return;
-        }
-
-        UpdateMoveSpeed(speedOverride);
-
-        float delta = (float) tickDelta / 1000;
-
-        if (walkType == WalkType.Direction) {
-            Vector3 newPosition = actor.Position + delta * Speed * walkDirection;
-            int searchRadius = Math.Max((int) (delta * Speed * 1.1f), 10);
-
-            actor.Navigation.UpdatePosition();
-            actor.Position = actor.Navigation.FindClosestPoint(newPosition, searchRadius);
-
-            Velocity = Speed * walkDirection;
-
-            UpdateDebugMarker(actor.Position, debugNpc, tickCount);
-            UpdateDebugMarker(actor.Navigation.GetAgentPosition(), debugAgent, tickCount);
-
-            return;
-        }
-
-        Vector3 offset = walkSegmentSet ? walkSegment.end - actor.Position : new Vector3(0, 0, 0);
-        float distanceSquared = offset.LengthSquared();
-        float tickDistance = delta * Speed;
-
-        if (!walkSegmentSet || distanceSquared < tickDistance * tickDistance) {
-            if (walkSegmentSet) {
-                actor.Position = walkSegment.end;
-
-                tickDistance -= (float) Math.Sqrt(distanceSquared);
-            }
-
-            walkSegment = actor.Navigation.Advance(TimeSpan.FromSeconds(0.25), Speed, out walkSegmentSet);
-
-            offset = walkSegmentSet ? walkSegment.end - actor.Position : new Vector3(0, 0, 0);
-            distanceSquared = offset.LengthSquared();
-
-            if (!walkSegmentSet || distanceSquared == 0) {
-                Vector3 walkTargetOffset = walkTargetPosition - actor.Position;
-                float walkTargetDistance = walkTargetOffset.LengthSquared();
-
-                if (walkLookWhenDone && walkTargetDistance > 0) {
-                    actor.Transform.LookTo(Vector3.Normalize(walkTargetOffset));
-                }
-
-                walkTask?.Cancel();
-
-                return;
-            }
-
-            float distance = (float) Math.Sqrt(distanceSquared);
-
-            walkDirection = (1 / distance) * offset;
-            tickDistance = Math.Min(tickDistance, distance);
-
-            actor.Transform.LookTo(walkDirection);
-        }
-
-        Velocity = Speed * walkDirection;
-        actor.Position += tickDistance * walkDirection;
-
-        Vector3 targetOffset = walkTargetPosition - actor.Position;
-        float targetDistance = targetOffset.LengthSquared();
-        float travelDistance = Speed * delta;
-
-        UpdateDebugMarker(actor.Position, debugNpc, tickCount);
-        UpdateDebugMarker(walkTargetPosition, debugTarget, tickCount);
-        UpdateDebugMarker(actor.Navigation.GetAgentPosition(), debugAgent, tickCount);
-
-        bool reached;
-
-        if (walkType == WalkType.MoveTo) {
-            reached = targetDistance < travelDistance * travelDistance;
-        } else if (walkType == WalkType.ToTarget) {
-            reached = targetDistance < walkTargetDistance * walkTargetDistance;
-        } else {
-            reached = targetDistance >= walkTargetDistance * walkTargetDistance;
-        }
-
-        if (reached) {
-            Velocity = new Vector3(0, 0, 0);
-
-            if (walkLookWhenDone) {
-                actor.Transform.LookTo(Vector3.Normalize(targetOffset));
-            }
-
-            walkTask?.Completed();
         }
     }
 
@@ -514,60 +324,4 @@ public partial class MovementState {
                 return false;
         }
     }
-
-    #region DebuggingImpl
-    private DebugMarker InitDebugMarker(int itemId, int rarity) {
-        actor.Field.ItemMetadata.TryGet(itemId, out ItemMetadata? itemData);
-
-        if (itemData is null) {
-            throw new InvalidDataException("bad item");
-        }
-
-        Item? item = actor.Field.ItemDrop.CreateItem(itemId, rarity);
-        if (item == null) {
-            throw new InvalidDataException("bad item");
-        }
-
-        FieldItem fieldItem = new FieldItem(actor.Field, 0, item) {
-            FixedPosition = true,
-            ReceiverId = -1,
-            Type = DropType.Player
-        };
-
-        return new DebugMarker(fieldItem, item);
-    }
-
-    private void UpdateDebugMarker(Vector3 position, DebugMarker marker, long tickCount) {
-        if (tickCount < marker.NextUpdate) {
-            return;
-        }
-
-        marker.NextUpdate = tickCount + 50;
-
-        if (marker.Spawned) {
-            //actor.Field.Broadcast(FieldPacket.RemoveItem(marker.ObjectId));
-        }
-
-        marker.Spawned = true;
-        //marker.ObjectId = actor.Field.RemoveMeNextLocalId();
-        //
-        //actor.Field.Broadcast(FieldPacket.DropDebugItem(marker.Item, marker.ObjectId, position, 1000000, 0, false));
-    }
-
-    private void RemoveDebugMarker(DebugMarker marker, long tickCount) {
-        if (tickCount < marker.NextUpdate) {
-            return;
-        }
-
-        if (!marker.Spawned) {
-            return;
-        }
-
-        marker.NextUpdate = tickCount + 50;
-
-        marker.Spawned = false;
-        //actor.Field.Broadcast(FieldPacket.RemoveItem(marker.ObjectId));
-        marker.ObjectId = 0;
-    }
-    #endregion
 }
