+++ Maple2.Server.Game/Model/Field/Actor/ActorStateComponent/MovementStateTasks/MovementState.WalkTask.cs
@@ -1,6 +1,5 @@
 ﻿using Maple2.Model.Enum;
 using Maple2.Server.Game.Model.Enum;
-using Maple2.Server.Game.Model.Skill;
 using System.Numerics;
 using static Maple2.Server.Game.Model.Field.Actor.ActorStateComponent.TaskState;
 
@@ -180,146 +179,5 @@ public partial class MovementState {
 
         UpdateMoveSpeed(speed);
         StartWalking(sequence, task);
-    }
-
-    public class NpcStandbyTask : NpcTask {
-        private MovementState movement;
-        public IActor? Target { get; init; }
-        public string Sequence { get; init; } = string.Empty;
-        public bool IsIdle { get; init; }
-        override public bool CancelOnInterrupt { get => true; }
-
-        public NpcStandbyTask(TaskState taskState, MovementState movement, string sequence, NpcTaskPriority priority, bool isIdle) : base(taskState, priority) {
-            this.movement = movement;
-        }
-
-        override protected void TaskResumed() {
-            movement.Standby(this, Target, IsIdle, Sequence);
-        }
-    }
-
-    private void Standby(NpcTask task, IActor? target, bool isIdle, string sequence) {
-        Idle(sequence);
-
-        if (target is null) {
-            return;
-        }
-
-        actor.Transform.LookTo(Vector3.Normalize(target.Position - actor.Position));
-    }
-
-    public class NpcEmoteTask : NpcTask {
-        private MovementState movement;
-        public string Sequence { get; init; } = string.Empty;
-        public bool IsIdle { get; init; }
-        override public bool CancelOnInterrupt { get => true; }
-
-        public NpcEmoteTask(TaskState taskState, MovementState movement, string sequence, NpcTaskPriority priority, bool isIdle) : base(taskState, priority) {
-            this.movement = movement;
-            Sequence = sequence;
-            IsIdle = isIdle;
-        }
-
-        override protected void TaskResumed() {
-            movement.Emote(this, Sequence, IsIdle);
-        }
-        protected override void TaskFinished(bool isCompleted) {
-            movement.Idle();
-        }
-    }
-
-    private void Emote(NpcTask task, string sequence, bool isIdle) {
-        if (!CanTransitionToState(ActorState.Emotion)) {
-            task.Cancel();
-
-            return;
-        }
-
-        if (!actor.AnimationState.TryPlaySequence(sequence, 1, AnimationType.Misc)) {
-            task.Cancel();
-
-            return;
-        }
-
-        ActorSubState subState = ActorSubState.EmotionIdle_Idle;
-
-        if (isIdle) {
-            subState = sequence switch {
-                "Bore_A" => ActorSubState.EmotionIdle_Bore_A,
-                "Bore_B" => ActorSubState.EmotionIdle_Bore_B,
-                "Bore_C" => ActorSubState.EmotionIdle_Bore_C,
-                _ => ActorSubState.EmotionIdle_Idle
-            };
-        }
-
-        emoteActionTask = task;
-        stateSequence = actor.AnimationState.PlayingSequence;
-
-        SetState(isIdle ? ActorState.Emotion : ActorState.EmotionIdle, subState);
-    }
-
-    class NpcSkillCastTask : NpcTask {
-        private MovementState movement;
-
-        public SkillRecord? Cast;
-        public int SkillId { get; init; }
-        public short SkillLevel { get; init; }
-        public long SkillUid { get; init; }
-
-        public NpcSkillCastTask(TaskState queue, MovementState movement, int id, short level, long uid) : base(queue, NpcTaskPriority.BattleAction) {
-            this.movement = movement;
-            SkillId = id;
-            SkillLevel = level;
-            SkillUid = uid;
-        }
-
-        override protected void TaskResumed() {
-            movement.SkillCast(this, SkillId, SkillLevel, SkillUid);
-        }
-
-        protected override void TaskFinished(bool isCompleted) {
-            movement.castSkill = null;
-            movement.castTask = null;
-            movement.Idle();
-            movement.actor.AppendDebugMessage((isCompleted ? "Finished" : "Canceled") + " cast\n");
-        }
-    }
-
-    private void SkillCast(NpcSkillCastTask task, int id, short level, long uid) {
-        castTask?.Cancel();
-
-        if (!CanTransitionToState(ActorState.PcSkill)) {
-            task.Cancel();
-
-            return;
-        }
-
-        Velocity = new Vector3(0, 0, 0);
-
-        SkillRecord? cast = actor.CastSkill(id, level, uid);
-
-        if (cast is null) {
-            return;
-        }
-
-        if (!actor.AnimationState.TryPlaySequence(cast.Motion.MotionProperty.SequenceName, cast.Motion.MotionProperty.SequenceSpeed, AnimationType.Skill)) {
-            task.Cancel();
-
-            return;
-        }
-
-        castTask = task;
-        castSkill = cast;
-        task.Cast = cast;
-
-        //if (faceTarget && actor.BattleState.Target is not null) {
-        //    actor.Transform.LookTo(Vector3.Normalize(actor.BattleState.Target.Position - actor.Position));
-        //}
-
-        SetState(ActorState.PcSkill, ActorSubState.Skill_Default);
-
-        stateSequence = actor.AnimationState.PlayingSequence;
-
-        return;
     }
 }
